requirements.txt
pytest==8.3.5
pytest-cov==6.1.1
fastapi>=0.68.1,<0.69.0
httpx>=0.27.0
python-multipart==0.0.5
pydantic>=1.9.0,<2.0.0
scikit-learn>=1.3.0
numpy>=1.26.0
PyMuPDF==1.23.8
python-docx>=0.8.11
langid>=1.1.6
textstat>=0.7.3
nltk>=3.8.1
openai==1.3.0
uvicorn==0.15.0
torch>=2.0.0
transformers>=4.30.2
PyPDF2>=3.0.1
pinecone-client>=2.2.1
python-dotenv==0.19.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==3.2.0
requests==2.26.0
aiofiles==0.7.0
python-magic==0.4.24
pandas>=2.0.0
tqdm>=4.65.0